requirements.txt
python-telegram-bot==21.10
aiosqlite==0.20.0
python-dotenv==1.0.1
